requirements.txt
# Django & REST
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
django-environ==0.11.2
django-extensions==3.2.3
python-decouple==3.8

# Database
psycopg2-binary==2.9.9
redis==5.1.0
django-redis==5.4.0

# ASGI / WSGI
uvicorn[standard]==0.35.0
gunicorn==23.0.0

# Celery
celery==5.5.3
django-celery-beat==2.8.1
django-celery-results==2.5.1
celery-progress==0.3

# Auth & Security
djangorestframework-simplejwt[crypto]==5.5.1
django-allauth==65.11.2
cryptography==42.0.2
django-ratelimit

# API Documentation
drf-spectacular==0.28.0

# Files / Storage
Pillow==11.3.0
django-storages==1.14.2
boto3==1.40.33

# Payments
stripe==12.5.1

# Monitoring
sentry-sdk[django]==2.38.0
prometheus-client==0.22.1
django-prometheus==2.3.1

# Search
django-haystack==3.2.1
elasticsearch==8.12.1

# Utils
requests==2.31.0
user-agents

# Development
django-debug-toolbar==4.2.0
factory-boy==3.3.0
faker==22.6.0
ruff==0.8.4
mypy==1.13.0
django-stubs==5.1.1

# Testing
pytest==8.4.2
pytest-django==4.11.1
pytest-cov==7.0.0
pytest-mock==3.12.0
